merging changes from frontend and data_service into master

+++ src/main/java/com/accenture/inteview/repository/QuestionRepository.java
@@ -13,6 +13,6 @@ public interface QuestionRepository extends JpaRepository<QuestionEntity, Long> 
 
 	@Modifying
 	@Query("DELETE FROM QuestionEntity q WHERE q.id = :id")
-	Long deleteQuestionById(@Param("id") Long id);
+	int deleteQuestionById(@Param("id") Long id);
 
 }
